# services/frontend/internal/backend/openapi.yml
openapi: 3.1.1
info:
  title: Mauerstrassenloewen API
  description: API for interacting with the Mauerstrassenloewen platform.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Reverse proxy server
  - url: http://localhost:8082
    description: Backend/msl service
  - url: http://localhost:8083
    description: Simulation service
paths:
...
  /msl/api/v1/chatbot/start-chat:
    post:
      summary: Start a new chat
      operationId: startChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartChatRequest'
      responses:
        '200':
          description: Chat started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        '400':
          description: Bad Request
        '415':
          description: Unsupported Media Type
        '500':
          description: Internal Server Error
...
components:
  schemas:
    StartChatRequest:
      type: object
      required:
        - system_msg
      properties:
        user_msg:
          type: string
          description: The user's initial message.
        system_msg:
          type: string
          description: The system message used to guide the assistant.
...
    Message:
      type: object
      properties:
        role:
          type: string
          description: The role of the message sender (e.g., user, system, assistant).
        content:
          type: string
          description: The message content.
        sent_at:
          type: string
          format: date-time
          description: Timestamp when the message was sent.
    Chat:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the chat session.
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: List of messages exchanged in the chat.
...